feat: pause and block queues
This change adds functionality to pause and block queues:

* Pausing stops dequeuing jobs, but they can continue to be enqueued,
  and will be executed after the queue is unpaused.
* Blocking disallows enqueuing new jobs.
* Any jobs enqueued before a queue is paused will continue to run.

Because jobs that were enqueued but haven't completed will continue to
run after a queue is paused, it is redundant to both pause and block a
queue. Blocking a queue will also functionally pause it.

Maybe a flag, --cancel, can be provided to cancel any in-progress jobs.