use bytes instead of interface{} for user data
This change changes the API to use `bytes` for args, data, results,
and checkin data, which generally simplifies the code. There's no point
in decoding JSON other than the basic validation (and requirement that
JSON is used, which there is no reason for). The code is also doing less
decoding and reencoding this way.